Some final changes
I'm not entirely sure what is being asked, as some of these are kind of vague. I'm not sure if it's meant to be vague, so if I am wrong with what I believe the answer should be or if I'm completely wrong, I will change it for the final portfolio.